requirements.txt
aiohttp==3.8.5
asyncpg==0.27.0
bcrypt==4.0.1
requests==2.31.0
PyJWT==2.8.0
pydantic==1.10.12
python-dotenv==1.0.1